Use a persistent volume for the remote local setup
By changing the remote local setup 'deployment' to a 'statefulset' we can use a
persistent volume for the home folder of the remote local setup pod.

The home folder contains valuable artifacts that need not be recreated each
time the remote local setup pod is started. These artifacts include:

- the git clone of the gardener/gardener repository,
- the golang compiler cache,
- the container images that were downloaded by
  - the dind docker daemon (e.g. the KinD image) or
  - containerd inside the KinD cluster or shoot nodes.

    These images are cached by the registry mirror pods in the dev folder of the
    gardener repository.

This allows us to persist this state across restarts and can significantly
reduce the startup time of the local setup scenario in subsequent executions:

- The golang compiler can reuse the compiled artifacts of the packages that did
  not change since the previous build,
- the previously used container images do not need to be downloaded by dockerd
  or containerd from the Internet and
- skaffold can reuse the previously built images and does not need to rebuild
  them.

This makes the remote local setup pod as efficient as a regular local setup
running on a developer's machine, where these artifacts are typically preserved.

To get a fresh state, developers can simply

- scale the remote local setup statefulset to zero replicas,
- delete the persistent volume claim and
- scale the statefulset back to one replica.

Getting a fresh state is typically not necessary as these artifacts are managed
by git/golang/registry and can be safely reused. They are not deleted on the
local machine either.

This change also allows developers to scale the remote local setup statefulset
to zero replicas when they do not need the remote local setup pod, which can
save cloud resources. This allows developers to share a single Gardener shoot
cluster to host multiple remote local setup pods in different namespaces.

- The cluster itself can be hibernated at night, and the individual remote local
  setup statefulsets can be scaled to zero replicas when the specific developer
  does not need the remote local setup pod. This way the cluster autoscaler can
  remove unnecessary nodes and save cloud resources.

- When the developers need the remote local setup pod again, they can scale the
  statefulset back to one replica and continue with their work where they left
  off with all the artifacts and caches still in place.