Add dependency k8s.io/kube-state-metrics/v2 v2.11.0
We can't upgrade to the latest k8s.io/kube-state-metrics/v2 due to a
version conflict with k8s.io/component-base v0.29.5.

The next version of k8s.io/kube-state-metrics/v2 requires a version of
github.com/prometheus/common (v0.51.1) that contains a breaking change.
This breaking change is incompatible with the version of
k8s.io/component-base (v0.29.5) that we're using.

By using k8s.io/kube-state-metrics/v2 v2.11.0 instead, we can avoid this
issue, as this version does not require the breaking change in
github.com/prometheus/common.

This change should not affect the runtime behavior, as the changes
between k8s.io/kube-state-metrics/v2 v2.11.0 and higher do not affect
the YAML serialization of the custom resource state data structures.

Once a new version of kubernetes/component-base is released containing
the fix
https://github.com/kubernetes/component-base/commit/62c04b35eff68f7d4087f94b7b1228facb375141
for the breaking change
https://github.com/prometheus/common/commit/bd41eb6b9dee4fa983f31ae8756700efde1f3ea2
that was introduced in v0.48.0 of prometheus/common, we can upgrade the
kube-state-metrics/v2 beyond v2.11.0.

Otherwise when using kube-state-metrics/v2 in the latest version the
build fails with a compilation error.

```
go get k8s.io/kube-state-metrics/v2@v2.11.0
```

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>