Show a greeting when exec-ing into the pod with 'k exec -it ... -- sh'
Previously the greeting 'To attach to the gardener tmux session...'
was shown only when bash was executed and it was even
shown inside the tmux session.

Now it is shown when the alpine default shell: ash is executed, so
when we exec into the pod with
  k exec -it remote-local-setup-0 -- sh
and it is not shown when already in a tmux session.

The text of the greeting is a bit more verbose.