Rewrite the alerts to check if there is quorum
The old alert of minimum 2 replicas running would always fire on non-HA
setups because in those cases there is only one instance of the etcd-main
or etcd-events ever deployed.

A quorum exists if there is a majority of etcd instances up and running
e.g., floor(N/2)+1, being N the number of etcd instances deployed.
That means, we can alert if there are less or equal than N/2 available
replicas as a simplification of the previous quorum formula. For example:

- N = 1 -> Alert if there are 1/2 (=0.5) or less healthy replicas, i.e.,
  do not tolerate a single replica down.
- N = 3 -> Alert if there are 3/2 (=1.5) or less healthy replicas, i.e.,
  tolerate one replica down.
- N = 4 -> Alert if there are 4/2 (=2) or less healthy replicas, i.e.,
  tolerate one replica down because 2 healthy replicas do not conform a
  majority out of 4 replicas. This example is however invalid because
  Gardener accepts only an odd number of etcd replicas, but it's used
  here for illustration.