Configure a simple, black and white skin for k9s
In the default MacOS Terminal app, with white background, the default
skin and the alternative skins do not look good / are not really
usable, e.g. the row selection color is too similar to the row text
color, or the prompt suggestions use a barely visible color.

This probably "Keep it simple ..." skin seems to be the best option,
even if not perfect in a white tmux session, e.g. the confirmation
dialog's OK/Cancel buttons are not highlighted when creating a
port-forward.