Increase the resource requests to require a 16 core / 32 GB node
A 16 core and 32 GB k8s node that is expected to run a single instance
of the remote local setup pod has comparable compute resources like a
modern developer laptop.

This capacity is sufficient to run the local setup in a smooth way.

Note that the resource requests are actually smaller than the expected
capacity of the node: 32GB -> 28GB, 16 core -> 14 core.
This reduction is needed to consider the kube reserved capacity: some
CPU and memory is dedicated for the kubelet and other system processes
and those resources are not available for regular pods.

In an example Gardener shoot with default configuration:

k describe node <node>
Capacity:
  cpu:                16
  ephemeral-storage:  64746412Ki
Allocatable:
  cpu:                15920m
  ephemeral-storage:  62985309545

Furthermore, it leaves some place to run daemonsets or other system pods
on the node. These system pods have typically a high priority such that
they would make room for themselves and the remote local setup pod
would be unschedulable if they can't all fit on a node.

The buffer of 2 cores and 4 GB is just an estimate. It is important
that the resource request is more than the size of a next smaller node
(8core/16GB) and that on a single node of the expected size (16core/32GB)
only 1 remote local setup pod would fit.

Note that using even larger nodes where multiple remote local setup pods
could fit is a configuration that I did not test: the recommendation is
to use one node for one remote local setup pod.