Use wait to keep the remote-local-setup pod running
Previously reading from stdin was used to block the init script.

The docker daemon is started as a shell background job so it is more
idiomatic to wait for that background job to finish.
The dockerd process is not going to finish so this similarly blocks
the init script indefinitely.

Using this approach, the PID 1 ash shell reaps the children that it
adopts from parents that exited and left some orphan child processes
behind, i.e. this approach prevents the accumulation of zombie processes.