Improve tmux behavior when 2 developers collaborate in the same session
When multiple developers attach to the same tmux session, use the smallest
window size.

The default behaviour is latest, such that tmux uses the size of the
client that had the most recent activity. This is convenient for the
active client, but confusing for the passive peer developer as tmux might
show only a fraction of the screen.

For collaboration it is beneficial to align the window size during the
call and using the smallest window size helps to make the window size
difference apparent.

http://man.openbsd.org/OpenBSD-current/man1/tmux.1