Add bash functions to support a PR test workflow
When testing a PR,
it is helpful to checkout the base commit of the PR first,
start the local setup, inspect its relevant state,
then checkout the head commit of the PR,
upgrade the local setup and
assert the expected change in the relevant state.

The pr and pr_base functions help to checkout the respective git commits.