Use a copy in the default kubeconfig path: ~/.kube/config
Previously, the KUBECONFIG was set to the path of the kubeconfig
that is generated by Gardener's make kind.*-up target.

This leads to issues when using the 'kns' tool: kns edits the kubeconfig
in place and sets the namespace of the current context accordingly.
The skaffold / helm based deployment of gardener fails if the
namespace is set in the kubeconfig that make gardener-up uses.

  make kind-up     # the namespace is not specified in the kubeconfig
  make gardener-up # OK
  kns kube-system  # specify a namespace in the kubeconfig
  make gardener-up # fails with
    waiting for deletion: ...
    You must pass '--namespace=garden' to perform this operation.
    the namespace from the provided object "garden"
    does not match the namespace "kube-system".

It is common to use kubectl alias-es like ks to work in the kube-system
namespace.

To allow to continue using the kns tool, the mitigation is to use
a copy of the Gardener generated kubeconfig in interactive sessions.
This way the copy can be freely edited without impacting the
Gardener deploy process. This is also mentioned in the documentation.

> > It might be helpful to copy this file to `$HOME/.kube/config`,
> > since you will need to target this KinD cluster multiple times.
> docs/deployment/getting_started_locally.md

The sync_kubeconfig function can be executed as a background bash job
in the tmux pane of 'make kind.*-up'. This is more convenient than
copying the regenerated kubeconfig after the kind-up target finished.
sync_kubeconfig uses inotify and copies the kubeconfig after it
changed.

If the sync_kubeconfig runs in the background, after
make kind-down kind-up, the newly generated kubeconfig is
automatically copied to ~/.kube/config