Tmux utility functions to facilitate preparing tmux sessions
When sending commands to bash in a tmux pane, we need to wait for bash
to be ready, i.e. to show the prompt.
Afterwards we need to wait for bash to echo back the command.
Otherwise the input can be sent too fast and some of the lines might get
lost (only) visually.