[release-v1.96] Fix etcd alerts for the virtual Garden cluster (#9974)
* Fix relabel config for etcd

The role should be obtained via the the __meta_kubernetes_pod_label_role
discovered label. The old code using
__meta_kubernetes_service_label_instance would produce a label such as
virtual-garden-etcd-main or virtual-garden-etcd-events.

We opt for fixing the label so that alerts and other queries do not need
to be fixed.

* Rewrite the alerts to check if there is quorum

The old alert of minimum 2 replicas running would always fire on non-HA
setups because in those cases there is only one instance of the etcd-main
or etcd-events ever deployed.

A quorum exists if there is a majority of etcd instances up and running
e.g., floor(N/2)+1, being N the number of etcd instances deployed.
That means, we can alert if there are less or equal than N/2 available
replicas as a simplification of the previous quorum formula. For example:

- N = 1 -> Alert if there are 1/2 (=0.5) or less healthy replicas, i.e.,
  do not tolerate a single replica down.
- N = 3 -> Alert if there are 3/2 (=1.5) or less healthy replicas, i.e.,
  tolerate one replica down.
- N = 4 -> Alert if there are 4/2 (=2) or less healthy replicas, i.e.,
  tolerate one replica down because 2 healthy replicas do not conform a
  majority out of 4 replicas. This example is however invalid because
  Gardener accepts only an odd number of etcd replicas, but it's used
  here for illustration.

* Fix indentation

* Fix unit tests

* Add alert unit test using PromQL

For both alerts EtcdMainDown and EtcdEventsDown, test:

- Etcd HA scenarios tolerate one replica down.
- Etcd non-HA scenarios alert as soon as the single instance is unhealthy.

* Remove new empty line from description

While we are on it, alert descriptions are not supposed to have empty
lines at the end because it's the alert template that is responsible for
formatting the final alert message.

* Fallback to using the old role label

There are some dashboards that would also need adaptation for this change
to work correctly. The change to roles with prefix "virtual-garden-etcd"
was intentional.

This reverts commit cc1a2ddd57b3194ef8e7249d193be1035f9bca4f.
This reverts commit 5f6c190546119b1ccac8e0ea96b93e5238f96d46.

---------

Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>