Fix etcd-druid alert
In 9f3b254, there was a regression in which the name of the etcd-druid
scrape job changed due to the switch to the prometheus operator. This led
to the `PrometheusCantScrape` alert being fired for the `etcd-druid` job.
This change excludes the `etc-druid` job from the alert again, as initially
intended.

To reiterate the motivation for excluding etcd-druid metrics: If there are
no compaction failures, no relevant metrics are exposed by the druid.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Christoph Kleineweber <c.kleineweber@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>