Add a progress indicator if the remote local setup is being initialised
The reinitialisation of the remote local setup pod takes about 40s.

  $ k delete pod remote-local-setup-0
  pod "remote-local-setup-0" deleted

  $ k get pod -w
  NAME                   READY   STATUS    RESTARTS   AGE
  remote-local-setup-0   0/1     Running   0          1s
  remote-local-setup-0   0/1     Running   0          40s
  remote-local-setup-0   1/1     Running   0          40s

If the developers exec into the pod earlier, and attach to the
tmux session, they might see how the tmux session is being prepared:
windows and panes are opened and the commands are entered, ...

During this time, the tmux session is actually not expected to be used.

This progress indicator now shows to the developers that the
initialisation process is not completed yet.

This improves the usability a bit - otherwise, developers would need to
wait with `k get pods -w` for the pod to become ready to avoid exec-ing
to the pod too early. Now they can exec into to the pod right away.

If something goes wrong, the otherwise blocking progress indicator
can be cancelled with C-c.