Add an example to create a managed seed
This is helpful to check the artifacts that are deployed to a managed
seed.

Note that the KinD cluster is conceptually the merge of the
runtime-garden cluster and a seed cluster, so a managed seed is useful
to check the artifacts of a seed, because it does not contain artifacts
of the runtime-garden cluster.