Fix the color issue of k9s inside tmux
In https://github.com/gardener/gardener/commit/728248f38d409b852aed12653669363e53bedb01
the line
  export TERM=xterm-256color
was added when k9s was installed.

In the commit of this PR: "tmux automatically sets the TERM env variable"
I removed this line.

Although it is true that tmux sets TERM to tmux-256color, but tmux itself
also needs to know that the terminal emulator supports 256 colors, when
it attaches to a session.

Furthermore, although not recommended, the developer could also start
bash outside of tmux: in this case, the TERM will be set to xterm-256color.
If bash is started inside tmux, it will be set by tmux to tmux-256color.

So we need to set TERM in the greeting script that sets the environment
variables for the ash shell.

Not exposing the terminal capabilities (256 colors) to tmux caused
the confusing situation that the colors were mapped to a smaller palette
by tmux and e.g. k9s with its default skin did not really look good.

With this proper color support even inside tmux, we can use the default
skin of k9s and it looks good.