Fix shoot:apiserver_latency_seconds:quantile federation configuration (#10009)
In 9f3b25415, the recording rule previously named
`apiserver_latency_seconds:quantile` was renamed to
`shoot:apiserver_latency_seconds:quantile`. This had two consequences. The
less significant is that it broke some dashboards. This commit fixes those
dashboards. The more significant one is that the aggregate prometheus then
started federating this recording rule due to the
`{__name__=~"shoot:(.+):(.+)"}` line in its federation parameters. This
expression is intended to federate metering metrics. Metrics matching that
expression are federated twice in gardener. Once, as just explained, from
the shoot control plane Prometheus to the aggregate Prometheus, and a
second time from the aggregate Prometheus to the garden Prometheus. By the
time this recording rule made it to the garden prometheus, it caused the
number of samples scraped to increase massively. This in turn was
significant enough for the garden Prometheus to consume more than twice as
much memory as before. Therefore, this commit also excludes this recording
rule from federation at the aggregate Prometheus level.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>