Fix VPA recommendations dashboards for the Prometheus custom resource
Since the introduction of the Prometheus operator as part of GEP-19, the
VPA targets the Prometheus custom resource which is called `shoot`. The
Prometheus pods are prefixed with "prometheus-", e.g.
`prometheus-shoot-0`, which breaks the VPA dashboard because the
assumption that the VPA target name is a prefix of the pod name is no
longer true: in this case the VPA target kind (`Prometheus`) is added as
a prefix to the pod name in lower case. So we use the `?i` flag to
expect the target kind to be an optional prefix of the pod name.