Remove use of externalLabels.landscape in alert from aggregate Prometheus
This label is only available in the garden Prometheus. The seed label
can be used instead despite it does not contain any information about
the landscape in the aggregate Prometheus.

However, this alert is federated into the garden Prometheus and
reimplemented there again to also be sent using the alert channels from
the garden Prometheus.

This commit adapts this alert description in the garden Prometheus to
add information about the seed and the landscape. This way, the alert
still does not contain information about the landscape when sent by the
seed Alertmanager, but it does now when sent by the garden Alertmanager.
Other alerts follow the same approach when federated into the garden
Prometheus.