Allow propagating network ranges from provider extension to Shoot (#9998)
* Extend `Infrastructure.Status` and `Shoot.Status` with a networking
section.

Especially, but not restricted to IPv6, infrastructures may provide the
actual cluster networking ranges similar to how it is already possible
today with regards to the node network. This change adapts the API to
allow specifying the ranges in infrastructure and propagating it to the
shoot status.

* Add validation for shoot networking status

* Adapt gardener extension library to consider value from the shoot
networking status in addition to the shoot specification.

* Extract networking status from infrastructure to shoot status

* Adapt gardenlet internal networking structs to allow multiple ranges per type

* Adapt kube-apiserver service

* Adapt kube-apiserver

* Adapt kube-controller-manager

* Adapt kube-proxy

* Adapt coredns

* Adapt nodelocaldns

* Adapt vpnseedserver

* Adapt network crd deployment

* Adapt additional botanist tests

* Adapt network policy controller

* Adapt gardenlet network policy controller predicate

* Adapt cluster autoscaler.

The calculation checking the sizes of pod and node range are now capable
of handling multiple ranges and support dual-stack networks. In
dual-stack networks, it is assumed that all pods/nodes get addresses
in both stacks. Therefore, the smaller network determines its maximum
size.

* Adapt hibernation check of endpoints

* Adjust seed/shoot network disjointness checks.

The shoot internal disjointness check was not adapted as with IPv6
it might not be useful anymore.

* Adapt provider-local to write Infrastructure.Status.Networking

* Address review feedback