Increase the max connections from 512->1024
Prometheus allows by default 512 incoming connections.

  docker run -it prom/prometheus --help | grep max-connections
  --web.max-connections=512  Maximum number of simultaneous connections.

Currently every shoot control plane Prometheus has 2 scrape jobs to federate
metrics from the cache Prometheus. So if there are 256 shoots in the seed, that
means 512 incoming connections. Prometheus and go in general maintains long
running TCP connections that are reused for multiple HTTP requests, so in such a
scenario the cache Prometheus has 512 long lived incoming connections.

With the default limit of 512, this means that other connections are not
accepted. This also affects the readiness and liveness probes that the kubelet
executes for the cache Prometheus, which are HTTP requests that are also counted
towards the limit of 512. A failing liveness probe has fatal consequences: the
kubelet will eventually restart the cache Prometheus. So in seeds with 256+
shoots, the cache Prometheus is constantly restarted by the kubelet.

The limit of 512 incoming connections is an arbitrary default value and is not
directly related to hard resource limits like CPU and memory.

With this commit, we increase the max connections limit by 2x, from 512 to 1024.
This should give enough headroom even in seeds with 256+ shoots.

To check the number of open connections in the cache Prometheus pod:

  k exec -it prometheus-cache-0 -- sh

  netstat -n | grep ESTABLISHED | grep :9090 \
  | awk '{print $5}' | sed -E 's/:[^:]+$//' | sort | uniq -c

      # prometheus-shoot, scales with the number of shoots in the seed
      2 ::ffff:10.1.101.160

      # prometheus-aggregate also scrapes the cache Prometheus,
      # but has only 1 replica
      1 ::ffff:10.1.14.29

      # prometheus-seed, also with 1 replica
      1 ::ffff:10.1.5.219

So the example above from the local setup shows that the number of incoming
connections scales linearly with the number of shoots in the seed. With the
increased limit of 1024, we should have enough headroom even in seeds with 256+
shoots.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Christoph Kleineweber <c.kleineweber@sap.com>