KSM server-side filter for allowed metrics
kube-state-metrics can be configured to expose only specific metrics
about the state of objects. In Gardener, we are not interested in
ingesting all available metrics into Prometheus but a subset of them.
Previously, this filtering was achieved via the scrape configuration on
the Prometheus side.

The kube-state-metrics in the seed is already configured to expose
certain metrics, but the kube-state-metrics in the shoot is not. Thus,
when the shoot Prometheus scrapes the kube-state-metrics exposed
metrics, it downloads a bunch of unwanted metrics and filters them out
via metric relabeling. This is inefficient because such filtering can be
done server-side to reduce the scrape time.

This commit improves this by extending the kube-state-metrics
configurations to make sure server-side filtering is in-place. This is
done via the `--metric-allowlist` flag, which expects a comma-separated
list of regular expressions. Metric names that match at least one of
those regular expressions are exposed. Therefore, metric names in
`--metric-allowlist` are now wrapped around the ^ and $ characters to
fully match the metric name.

The client-side filtering from the Prometheus scrape configuration is
now removed as it is not necessary. The kube-state-metrics deployments
expose only the desired metrics and Prometheus injects everything that
is scraped. Keeping the client-side filtering could be seen as a second
line of defense, but it would also add a bit of configuration overhead
without a clear benefit.

As for the garden cluster, the kube-state-metrics is also deployed with
its specific allowed metric configuration. Previously, it was deployed
as for a regular seed and undesired metrics were filtered out by a
scrape config. VPA metrics for the garden cluster are also added to the
garden Prometheus as a preparation for further work to alert on certain
VPA recommendations, including resources in the garden namespace. In
consequence, the number of samples in the garden Prometheus increases.
There are around 250 containers deployed in the garden cluster and 13
VPA metrics per container. This adds up to 250*13=3250 new samples. This
is just a rough maximum estimation as not all containers are monitored
by VPA, so the real increase is probably somewhat lower. A side-effect
of ingesting metrics from the garden namespace is that the number of
samples for the kube_pod_container_status_restarts_total and
kube_pod_status_phase metrics also increases. However, this is also not
significant. Such metric cardinalities are five and one, respectively:
there are five pod states (pending, succeeded, failing, unknown and
running) and only one restart counter per pod. The garden cluster
deploys around 110 pods in the garden namespace. Consequently, this
increases the number of samples up to 110*(5+1)=660.