Add containerd config in `OperatingSystemConfig` (I) (#10050)
* Add `containerd` to OSC API

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Add `EnsureCRIConfig` to CP ensurer

Extensions should be able to adjust the CRI config via the CP webhook, for example to add registry mirrors.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Copy `criConfig` to GNA secret

`gardener-node-agent` (GNA) needs the `criConfig` in the OSC reconciler for creating the desired containerd configuration. See the following commit(s).

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Handle registry configuration

This commits add the registry handling to `gardener-node-agent`. Along the way, the creation of required containerd directories was moved from the `containerd-initializer` service to the OSC controller.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Move `containerd` default config

The default config can be written by `gardener-node-agent` instead.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Move `containerd` service environment

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Move registry and import paths configuration

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Handle CRI != `containerd` case

This case is rather theoretical today, since we only allow and default to `containerd`. It serves as an additional safety net in the future.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Move `SandboxImage` handling

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Prepare unit deletion

This commit starts the `containerd.service` via `gardener-node-agent`. If containerd is already running, the command is solely a no-op.

Gardener extensions may define dependencies on `containerd` via a Dropin file. (e.g. https://github.com/gardener/gardener-extension-os-gardenlinux/blob/a2debf66f59d4d163f53fbfcf6fe3b0911ec3805/pkg/controller/operatingsystemconfig/actuator.go#L81). A deletion of such services leads to a halt of `containerd`.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Prune `containerd-initializer` service

The service is not relevant anymore, but we can't delete it yet on existing nodes because it causes a halt of `containerd` (see previous commit).

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Remove `containerd-initializer` for new nodes

On new nodes, there is no earlier defined dropin
on containerd. The deprecated `containerd-initializer`
can be skipped safely.

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Handle local setup

This commit adds possible registries via the `EnsureCRIConfig` functionality in the Local Provider.
A comment about the Registry-Cache extension is removed because it is outdated. The extension doesn't add `hosts.toml` files to the OSC directly (see https://github.com/gardener/gardener-extension-registry-cache/blob/main/pkg/webhook/cache/ensurer.go).

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* Enhance integration test

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>

* [make tidy]

* Relax `criConfig` validation

The validation in place must be removed to allow
shoot deletion to continue, that was already triggered
with an earlier Gardener version. Otherwise, the following
error will happen:

```
2024-07-02T09:36:04.359885376Z stderr F {"level":"error","ts":"2024-07-02T09:36:04.359Z","msg":"Error","controller":"shoot","object":{"name":"e2e-upgrade","namespace":"garden-local"},"namespace":"garden-local","name":"e2e-upgrade","reconcileID":"12c3d8f7-8d18-4717-987b-7e27fb915d2d","operation":"delete","flow":"Shoot cluster deletion","task":"Deleting operating system config resources","error":"retry failed with context deadline exceeded, last error: 1 error occurred:\n\t* admission webhook \"validation.extensions.operatingsystemconfigs.resources.gardener.cloud\" denied the request: OperatingSystemConfig.extensions.gardener.cloud \"gardener-node-agent-local-ee46034b8269353b-original\" is invalid: spec.criConfig.containerd: Required value: containerd config is required when CRI is configured\n\n","stacktrace":"github.com/gardener/gardener/pkg/utils/flow.(*execution).runNode.func2\n\tgithub.com/gardener/gardener/pkg/utils/flow/flow.go:231"}
```

* Address review feedback

* Fix cluster creation with enabled readiness probes

For new clusters, where a registry cache/mirror
runs as workload in the cluster, failing readiness
probes block the process, because nodes don't turn
ready if the registry handling happens too early.

* Refine registry configuration

Registries with enabled readiness probes may need
running workload in the cluster, which is why they
should be added asynchronously.
Other registries, however, may be required on the
node very early during bootstrapping which
is why they should be added as soon as possible.

* Address review comments

* Retain order of registry configuration

Toml encoding can't be used to retain the host order because map iterations on Go are not stable and custom sorting functions can't be passed.

* Fix auto-merge

---------

Co-authored-by: Gerrit Schwerthelm <gerrit.schwerthelm@x-cellent.com>
Co-authored-by: Thomas Buchner <thomas.buchner@sap.com>