Wait for previous managed resource to be healthy
The cluster role and cluster role binding were originally managed by the
shoot-core-kube-state-metrics managed resource but are now managed by
the shoot-core-kube-state-metrics-target managed resource.

In order to migrate those two resources to the latter managed resource,
we need to wait until the former stops managing them and deletes them.
Otherwise, the previous managed resource wants to delete them while the
new one keeps creating them.

As for the tests, a second goroutine is introduced to watch to the
shoot-core-kube-state-metrics managed resource and set its generation
and status conditions to healthy.