Allow seed deletion in combined garden runtime cluster and soil cluster setup (#10143)
* Allow seed deletion in combined garden runtime cluster and soil cluster
setup.

Two managed resources included the same resource. One of them was
managed by gardener operator, the other one was in gardenlet's
responsibility. If gardener operator was deployed to a cluster that
later became a seed, deletion of the seed was ending in a race between
both gardener resource managers fighting over the single resource
(deleting/recreating them in the process).
As the resource is not useful at all in the garden runtime cluster, it
will only be deployed into seed clusters going forward.

* Address review feedback