Use CoreDNS server as DNS resolver for tests in operator-seed e2e tests
The test creates a non-HA shoot on a multi-zone seed. This results in
the fact that not the "global" Istio instance (in the `istio-ingress`
namespace) is exposing the shoot, but one of the zone-specific Istio
instances.
In order to resolve the shoot DNS name to the correct Istio instance, we
have to consult the CoreDNS server deployed as part of the
provider-local extension.