Prevent unrequired kind cluster configuration in operator setup
We don't need the ETCD node port mapping (`gardener-operator` uses
`etcd-druid` to manage the ETCD clusters which doesn't configure this
`32379` node port at all).

We don't want the kind `kube-apiserver` to get configured with the
authz webhook for the SeedAuthorizer feature (`gardener-operator`
deploys a dedicated `kube-apiserver` for the virtual garden cluster
which is configured appropriately).