[operator] Reconcile `seedmanagement.gardener.cloud/v1alpha1.Gardenlet` object when there is no corresponding `Seed` yet (#10161)
* Leftover in docs

after https://github.com/gardener/gardener/pull/10162

* Documentation

* Add `kubeconfigSecretRef` to `Gardenlet` API

* Prepare config API for new controller

* Move `ManagedSeed` actuator and values helper to common package

`gardener-operator` will reuse this logic in the future, so let's move
it to some common package.
In the next commit, we will make it generic by removing the
ManagedSeed-specifics from it.

* Prepare `Actuator` for reuseability

In the next commit, we will remove `Shoot`-specifics from it so that it
can also be used for the `gardener-operator` use case.

* Remove `Shoot`-specifics from `Actuator`

and move them into `ManagedSeed` controller

* Add `Gardenlet` controller to `gardener-operator`

It is only responsible when the `Seed` does not exist yet, i.e., only
for the very first deployment of `gardenlet`.

Deletion is not implemented since this is a human operator concern, just
like today.

* Integration test

* Adapt `operator-seed` local setup

to deploy `Gardenlet` resource instead of Helm chart

* Add missing port mapping in local operator setup when used as seed

- The kind cluster in the operator setup is "HA", i.e., it has three
  nodes. Previously, the gardenlet was wrongly deployed since its `Seed`
  defintiion (in its component config) was reused from the regular
  "gardener kind setup" which only has one node:
  - https://github.com/gardener/gardener/blob/426f1138a56bcd41dfc9ada93e1325d91c76786c/example/gardener-local/gardenlet/values.yaml#L78-L79
    vs.
    https://github.com/gardener/gardener/pull/10161/files#diff-3948c9f4b4621516f0d10ed59db5f225ae88a3217b3b225776cd107301071d19R25-R28

- This led to issues when trying to communicate with the API servers of
  shoot clusters hosted there in case they are not HA. In such cases,
  they are not exposed via the "global" istio, but via a zone-specific
  instance of istio. However, these zone-specific istios did not have
  a port mapping to the host machine, hence, it was impossible to reach
  them.
  - https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/10161/pull-gardener-e2e-kind-operator-seed/1815677484994138112

- Now, the port mappings are properly established.

* Prevent unrequired kind cluster configuration in operator setup

We don't need the ETCD node port mapping (`gardener-operator` uses
`etcd-druid` to manage the ETCD clusters which doesn't configure this
`32379` node port at all).

We don't want the kind `kube-apiserver` to get configured with the
authz webhook for the SeedAuthorizer feature (`gardener-operator`
deploys a dedicated `kube-apiserver` for the virtual garden cluster
which is configured appropriately).

* Use CoreDNS server as DNS resolver for tests in operator-seed e2e tests

The test creates a non-HA shoot on a multi-zone seed. This results in
the fact that not the "global" Istio instance (in the `istio-ingress`
namespace) is exposing the shoot, but one of the zone-specific Istio
instances.
In order to resolve the shoot DNS name to the correct Istio instance, we
have to consult the CoreDNS server deployed as part of the
provider-local extension.

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback