Fix and improve alert descriptions (#10195)
* Remove empty lines from alert descriptions

Alert descriptions and summaries should not end with empty lines because
they might break the final alert format generated by the alert template.

* Fix SeedNodeNotHealthy alert description

Commit b48c204 accidentally removed one line too much and broke the
SeedNodeNotHealthy alert description.

* Remove use of externalLabels.landscape in alert from aggregate Prometheus

This label is only available in the garden Prometheus. The seed label
can be used instead despite it does not contain any information about
the landscape in the aggregate Prometheus.

However, this alert is federated into the garden Prometheus and
reimplemented there again to also be sent using the alert channels from
the garden Prometheus.

This commit adapts this alert description in the garden Prometheus to
add information about the seed and the landscape. This way, the alert
still does not contain information about the landscape when sent by the
seed Alertmanager, but it does now when sent by the garden Alertmanager.
Other alerts follow the same approach when federated into the garden
Prometheus.

* Generalize SeedNodeNotHealthy alert message

This alert is federated from another alert in the aggregate Prometheus
in the seed. This was changed from five to ten scrapes but we missed to
adapt it in the garden Prometheus as well.

It is better to have a more generic message that matches possible future
changes.