Allow `machine-controller-manager-provider-local` to delete VMs in the safety controller. (#10222)
* Allow `machine-controller-manager-provider-local` to delete VMs in the
safety controller.

In #10176, the first step was taken to allow provider-local to delete
virtual machines in machine-controller-manager's safety controller.
Unfortunately, while the deletion of the user data secret succeeded, the
deletion of the pod itself still failed.
This change is a simple addon to the previous pull request and should
finally resolve this issue.

* Address review feedback