Fix `provider-local`'s machine-controller-manager driver implementation with regards to listing of machines.
The machine-controller-manager assumes to get a map of provider IDs to
machine names when listing machines. However, `provider-local`'s
implementation had the subtle bug that it returned a map of provider IDs
to pod/node names instead. The resulting values had an additional
'machine-' prefix.
This caused the safety controller in machine-controller-manager to act
incorrectly, trying to delete machines and failing to do so.
In #10176 and #10222, some of the existing issues were addressed, but
now it became obvious that the real issue is that `provider-local`'s
implementation did not conform to the expected interface.