[operator] Auto-update Helm chart reference in `Gardenlet`s when opted in (#10218)
* Adapt image vector to accept full `ref` of image

instead of separating it via `repository` and `tag`

* Rename `images.yaml` to `containers.yaml`

In a subsequent commit, we will introduce `charts.yaml` and a separate
image vector for it. Hence, let's use specific file names.

* Rename `imagevector.ImageVector()` to `Containers()`

again, to make it more specific

* Prefix image name Golang constants with `Container`

We might have images in the to-be-introduced charts image vector with
the same names, hence, the generated Golang constants should have unique
names. For charts, the prefix will be `Chart`, i.e., resulting Golang
constants will be sth like `ChartImageNameGardenlet`.

* Introduce Helm chart image vector

- Only `gardener-operator` uses it, hence, we only add support for
  overwriting it for this component.
- On the way, the mount paths were streamlined (`charts_overwrite`
  sounds confusing nowadays).

* ------------ empty separator commit -------------

* Adapt documentation

* Adapt `Garden` reconciler

* Adapt integration test

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback

* Keep generate-imagename-constants.sh backwards-compatible

* Address PR review feedback